pyproject.toml
[project]
name = "mtia"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "click",
    "sqlalchemy>=2.0",
    "psycopg2-binary",
    "pandas>=2.0",
    "lightgbm",
    "scikit-learn",
    "joblib",
    "pyarrow",
    "fastapi",
    "uvicorn[standard]",
    # RAG stack
    "httpx>=0.27",
    "qdrant-client>=1.12",
    "pypdf>=4.0",
    "pyjwt>=2.8",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.1",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "reportlab>=4.0",  # test-only: synthesize PDFs in ingest tests
]

[project.scripts]
stopreason = "modules.stopreason.pipeline:cli"

[tool.setuptools]
py-modules = ["api"]
packages = ["modules", "modules.stopreason", "modules.rag"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["modules"]
python_files = ["test_*.py"]
